Update FitSend.py
REQUIRES FFMPEG AND ITS ASSOCIATED LIBRARIES. These files must reside in the same location as the script. You also, obviously, must have the video in the same location. Currently, the script converts the video to images and saves them to the hard drive. Another process then performs a directory listing and begins packetizing the images and sending them via UDP across the network to the client. Statistics have been added to the end to demonstrate functionality.